+++ src/main/kotlin/cloud/glitchdev/rfu/utils/rendering/Shape.kt
@@ -1,5 +1,5 @@
 package cloud.glitchdev.rfu.utils.rendering
 
 enum class Shape {
-    SPHERE, CYLINDER, LINE
+    SPHERE, CYLINDER, LINE, TEXT
 }